e2e: revert snake_case renames to keep PR diff focused
